+++ wazo_phoned/plugins/yealink/bus_consume.py
@@ -20,6 +20,9 @@ class BusEventHandler:
             self._users_forwards_unconditional_updated,
         )
         bus_consumer.on_event(
+            'users_forwards_noanswer_updated', self._users_forwards_noanswer_updated
+        )
+        bus_consumer.on_event(
             'users_forwards_busy_updated', self._users_forwards_busy_updated
         )
 
@@ -33,6 +36,12 @@ class BusEventHandler:
         enabled = event['enabled']
         destination = event['destination']
         self._service.notify_forward_unconditional(user_uuid, destination, enabled)
+
+    def _users_forwards_noanswer_updated(self, event):
+        user_uuid = event.get('user_uuid') or event.get('user_id')
+        enabled = event['enabled']
+        destination = event['destination']
+        self._service.notify_forward_noanswer(user_uuid, destination, enabled)
 
     def _users_forwards_busy_updated(self, event):
         user_uuid = event.get('user_uuid') or event.get('user_id')