+++ wazo_phoned/plugins/yealink/plugin.py
@@ -14,6 +14,7 @@ from .http import (
     BusyForwardUserService,
     DNDUserService,
     Lookup,
+    NoAnswerForwardUserService,
     UnconditionalForwardUserService,
 )
 from .services import YealinkService
@@ -27,6 +28,7 @@ class Plugin:
 
     user_service_dnd_url_fmt = '/{vendor}/user_service/dnd'
     user_service_forward_busy_url_fmt = '/{vendor}/user_service/forward_busy'
+    user_service_forward_noanswer_url_fmt = '/{vendor}/user_service/forward_noanswer'
     user_service_forward_unconditional_url_fmt = '/{vendor}/user_service/forward_unconditional'
 
     vendor = 'yealink'
@@ -67,6 +69,7 @@ class Plugin:
 
         self.user_service_dnd_url = self.user_service_dnd_url_fmt.format(vendor=self.vendor)
         self.user_service_forward_busy_url = self.user_service_forward_busy_url_fmt.format(vendor=self.vendor)
+        self.user_service_forward_noanswer_url = self.user_service_forward_noanswer_url_fmt.format(vendor=self.vendor)
         self.user_service_forward_unconditional_url = self.user_service_forward_unconditional_url_fmt.format(vendor=self.vendor)
         self._add_resources(api, directories_class_kwargs)
         self._add_user_service_resources(api, user_service_class_kwargs)
@@ -90,6 +93,12 @@ class Plugin:
             BusyForwardUserService,
             self.user_service_forward_busy_url,
             endpoint='yealink_user_service_forward_busy',
+            resource_class_kwargs=class_kwargs,
+        )
+        api.add_resource(
+            NoAnswerForwardUserService,
+            self.user_service_forward_noanswer_url,
+            endpoint='yealink_user_service_forward_noanswer',
             resource_class_kwargs=class_kwargs,
         )
         api.add_resource(