+++ wazo_phoned/plugins/yealink/tests/test_services.py
@@ -202,3 +202,45 @@ class TestServices(unittest.TestCase):
                 ]
             }
         )
+
+    def test_forward_noanswer_notify_disable(self):
+        self.confd.users.get.return_value = {'lines': [{'name': 'line-123'}]}
+        self.service.notify_forward_noanswer('123', '1002', False)
+        self.amid.action.assert_called_once_with(
+            'PJSIPNotify',
+            {
+                'Endpoint': 'line-123',
+                'Variable': [
+                    'Content-Type=message/sipfrag',
+                    'Event=ACTION-URI',
+                    'Content=key=NoAnswFwdOff',
+                ]
+            }
+        )
+
+    def test_forward_noanswer_notify_errors(self):
+        http_error = HTTPError()
+        http_error.response = MagicMock()
+        http_error.response.status_code = 404
+        self.confd.users.get.side_effect = http_error
+        assert_that(
+            calling(self.service.notify_forward_noanswer).with_args('123', '1002', True),
+            raises(NoSuchUser),
+        )
+        http_error.response.status_code = 500
+        assert_that(
+            calling(self.service.notify_forward_noanswer).with_args('123', '1002', True),
+            raises(HTTPError),
+        )
+
+    def test_forward_noanswer_enable(self):
+        self.service.update_forward_noanswer('123', '1002', True)
+        self.confd.users('123').update_forward.assert_called_once_with(
+            'noanswer', {'destination': '1002', 'enabled': True}
+        )
+
+    def test_forward_noanswer_disable(self):
+        self.service.update_forward_noanswer('123', '1002', False)
+        self.confd.users('123').update_forward.assert_called_once_with(
+            'noanswer', {'destination': '1002', 'enabled': False}
+        )