+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
   * GET `/0.1/yealink/user_service/dnd`
   * GET `/0.1/yealink/user_service/forward_busy`
   * GET `/0.1/yealink/user_service/forward_unconditional`
+  * GET `/0.1/yealink/user_service/forward_noanswer`
 
 20.03
 -----